requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
black==24.1.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
cryptography==46.0.3
distlib==0.4.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
fastapi==0.122.0
filelock==3.19.1
flake8==6.1.0
Flask==2.2.5
gitdb==4.0.12
GitPython==3.1.45
google-adk==1.3.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.128.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-speech==2.34.0
google-cloud-storage==2.19.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.47.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==5.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.6.0
mypy_extensions==1.1.0
narwhals==2.12.0
nodeenv==1.9.1
numpy==2.0.2
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1

pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==3.3.1
proto-plus==1.26.1
protobuf==6.33.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.1.0
pyparsing==3.2.5
pytest==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.1
rsa==4.9.1
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
